+++ Forms/eac3toForm.vb
@@ -796,8 +796,8 @@ Public Class eac3toForm
 
                     If ms.Codec.Contains(",") Then ms.Codec = ms.Codec.Left(",")
 
-                    ms.IsVideo = ms.Codec.EqualsAny("h264/AVC", "VC-1", "MPEG2")
-                    ms.IsAudio = ms.Codec.EqualsAny("DTS Master Audio", "DTS", "DTS-ES", "DTS Hi-Res", "DTS Express", "AC3", "AC3 EX", "AC3 Headphone", "AC3 Surround", "EAC3", "E-AC3", "E-AC3 EX", "E-AC3 Surround", "TrueHD/AC3", "TrueHD/AC3 (Atmos)", "TrueHD (Atmos)", "RAW/PCM", "MP2")
+                    ms.IsVideo = ms.Codec.EqualsAny("h264/AVC", "h265/HEVC", "VC-1", "MPEG2")
+                    ms.IsAudio = ms.Codec.EqualsAny("DTS Master Audio", "DTS", "DTS-ES", "DTS Hi-Res", "DTS Express", "AC3", "AC3 EX", "AC3 Headphone", "AC3 Surround", "EAC3", "E-AC3", "E-AC3 EX", "E-AC3 Surround", "TrueHD/AC3", "TrueHD/AC3 (Atmos)", "TrueHD (Atmos)", "RAW/PCM", "MP2", "AAC")
                     ms.IsSubtitle = ms.Codec.StartsWith("Subtitle")
                     ms.IsChapters = ms.Codec.StartsWith("Chapters")
 
@@ -976,6 +976,9 @@ Public Class eac3toForm
             Case "h264/AVC"
                 cbVideoOutput.Items.Add("H264")
                 cbVideoOutput.Items.Add("MKV")
+                cbVideoOutput.Text = If(M2TSFile = "", "MKV", "Nothing")
+            Case "h265/HEVC"
+                cbVideoOutput.Items.Add("H265")
                 cbVideoOutput.Text = If(M2TSFile = "", "MKV", "Nothing")
             Case "VC-1"
                 cbVideoOutput.Items.Add("MKV")