+++ My Project/AssemblyInfo.vb
@@ -7,5 +7,5 @@ Imports System.Runtime.InteropServices
 <Assembly: AssemblyCopyright("stax76")>
 <Assembly: ComVisible(False)> 
 <Assembly: Guid("54d5dc0f-6ba7-42f0-a82f-9c71d561cb8b")>
-<Assembly: AssemblyVersion("1.7.0.3")>
-<Assembly: AssemblyFileVersion("1.7.0.3")>
+<Assembly: AssemblyVersion("1.7.0.4")>
+<Assembly: AssemblyFileVersion("1.7.0.4")>